\name{auth}
\alias{auth}
\title{Auth user.}
\usage{
auth(login, password)
}
\arguments{
  \item{login}{Login user to auth.}

  \item{password}{Password to auth user in the system.}
}
\description{
Auth user to system using pam module
}
\details{
This function is used to auth user to system using pam
module.
}

